[DEV] Disable possible tap cop
Seems like it is impossible to correctly define where somebody can use
tap because of huge variaty of possible usings and meanings.
Additionally it's hard to work with early returns inside tap and so on.
I need to disable it because have no better idea how to make it more
stable.